Quitting darts whilst in the game now possible using pause menu and main menu button in this menu.
Pressing main menu now whilst within a game will not do anything if in a main game.
If however we are in a darts game then this game gets cancelled and we are returned to the main menu.

IM GOING TO RECORD MORE GAMEPLAY TOMORROW THAT IS HOPEFULLY BETTER YAYAY.