External unhovering and unselecting cards now actually unhovers or unselects them as well so the timelines are reset instead of just removing from variables/array.
Clear selected cards function and hover first card function in the current hand component now also calls the corresponding function in the card interface so that the timelines will get reset.
This should hopefully get rid of the glitchy instantaneous movement of cards in the hand when hovering and selecting after certain actions within games.