Joker Appearance fix. Can Select Card extra check for hand empty. Card lighting channel change. NPC audio volume change. Turned debug off for widget interaction component in player.
Joker had the wrong static mesh so was never using the correct texture and so we had 2 conquests but never saw the joker static mesh since it was named slightly differently but this is changed now so the joker has its proper appearance.

More checks for the can select card function for if the hand is empty same for the hover checks which should hopefully avoid infinite loops when having no cards in the hand.

Card lighting channel changed for main card actor and the holo card actor to just channel 1 so only the main area lights and not the spotlights.

Turned the debug off for the widget interaction component in the player for now can always turn it back on again later.