Toggle Keybinds.
New input action for toggling the button prompts.
New interface for the button prompt widgets.
Only one function in the interface for toggling the keybinds.
New enum for the main menu button prompt states.
New variable in the bar button prompt widget of type this new enum.
Updating the value of this variable in the different functions.
Using the enum variable for showing the button prompts again so they are not all overlapping.

Input action in player calling toggle keybinds function in the game mode.
New variable in the game mode for if the keybinds are visible.
Using current game enum to toggle the keybinds visibility.