Tuna animation setup. Tuna looks at player and possibly other characters now as well.
New animation blueprint for the cat.
Animation setup for the cat.
New enum for cat states.
New variables in the leading cat actor and the teleporting cat actor for if should look at character and what character to look at based on all character names enum.
Property access for these variables in the thread safe update.
Similar skeletal look at nodes for the cat looking but had to invert the axis.
Space and functionality setup for the moving animation for the leading cat and for the stationary animations for the cat in the animation blueprint.