Character Specific Action method overhaul. Hide Rules on first turn bug fixed for all games. Unique Card Indicator overhaul. Doubt call overhaul.
New interface function for the NPCs for the character specific functionality.
New enum for possible character specific actions.
Removed previous character specific action functionality.

New bools in all game managers except the pool one for now.
This checks if the game has been started
Use this rather than the current round index in the game mode for unpausing the game after closing the rules so that the game doesn't repeatedly start after hiding rules on the first turn of the game.
This bool is set back to the false in the end game function in each game manager except the pool one.

Unique card indicator widget changed to use a wrap box. New widget created for the text of the unique card indicator which is added to the wrap box.
Fade unique card indicator events now use the new pop-up animation in the widget instead of fading out so that there is some animation.
Storing timer of this fading in the game managers: auction and IDI. This timer is cleared when we update the unique card indicator so that there's no glitchness.
Size of font changes based on how many unique card indicator texts being displayed.

Doubt call widget changed to use pop-up animation for showing and for fading out.
No longer fading the colour and instead using this new animation.
So now theres at least some feedback for the doubt fading since it needs to be masked to look good with the post process.

Removed unused card indicator widgets.