More specificity for drink ordering. New drink order index variable in NPC and animation blueprint.
New variable in the NPC actor and the animation blueprint for NPCs.
This is the drink order index so can be more precise with which animation to play for which drink being ordered if there are going to be multiple animations.
This will also be used later when adapting the animation accessories for the drink making and giving so that we can choose specifically which mesh to use etc.
This variable is only ever going to be set from the drink order menu widget so no need to create new events or anything and instead we are just directly setting it since we have a direct reference to the dante npc in the widget.