Checked all dialogue calls. Changed dialogue link anim notify. Checked animation calls. Character specific calls in the drink order widget for make drink. Fixed Jukebox pausing and unpausing. Elongated fade times for pausing/unpausing. Final word more functionality implemented.
Checked through all of the dialogue calls that are currently in the functionality and think are all correct.
In the dialogue link anim notify added another variable for the specific array index of the dialogue that is supposed to play and then also turned off random from array in the dialogue call so more specific and will actually allow for proper chains of dialogue theoretically (still can't test since not been given lip sync animations for anyone else yet).

Think all the animation call / state updates are fine.

Currently only one set of do character specific action function calls which are now in the drink order widget for make drink.
Already setup the function so it should all be working.
Character specific actions are not for when in-game so that needs to be remembered.
The animations that supposedly are character specific will most likely be idle animations when in-game so don't need to do character specific calls in game managers probably.

Jukebox wasn't actually resuming when clicking pause when already paused so fixed this now.

More functionality for the entire game completing.
Now remove the button prompts by toggling visibility.
Player input restricted to UI only and can't look or move.
Only set mouse cursor visible after fade to black timer call.
New function for showing the end screen.
This is called after starting the fade to black currently with a 4 second timer.

New functionality in the game completed widget for resetting the player.
Allowing them to move again by setting up their inputs again.
New widget animations in this widget as well for hovering and clicking the 2 buttons.