New anim notify for combining accessories. New functionality for anim accessories. New smaller shaker model. All functionality for make drink animation done for now. Still needs some slight positioning tweaking for sockets. No longer activating head tracking when entering interact collision if doing a character specific action.
New Sockets on the skeleton.

New anim notifies: combining accessories, detatching accessories.
New system for accessories being stored in the world.
Spawning static mesh actors and then adding them to an array in the NPC that is responsible for them.
Multiple interface functions for manipulating these and the array they are stored in.
New functionality in the previously made anim notifies specifically for this animation.
New entries in the animation accessories enum.
Last thing to do for this is the tweaks and making the drink glass actually match the drink you order.

4 main notifies used for this animation are as follows: accessory equip, accessory unequip, attach accessories together, detach accessories from each other.
So all new functionality is in these notifies some of which are new entirely.

Also need to add functionality for the NPCs removing their accessories and destroying their placed accessories when they are forcibly entered a new state like in-game so that the accessories aren't still in their hands or when they were placed. Already set it up for resetting the placed ones when they enter the character specific action state.