Changed blend times again back down a bit for majority of the states. Changed where we set in game active in the game managers for NPCs.
Going to simplify a lot of the stuff ive done today for that drink animation since just realised an easier way to do it. Then will do the whole interrupting animations hiding the animation accessories thing.

Now the calls for setting in game for the NPC opponents is done in the setup functions for the game managers instead of in the begin game function so that their animations change sooner.