Updated biel and zilia models. Zilia needs cloth sim done on new version of model. Moved zilia and biel bar positions. Changed NPC hand splines. Fixed issue with starting every single main game.
Since updated Zilia model cloth sim needs to be done again on the dress.

Moved the positions at the bar for zilia and biel to be closer to the bar.
Changed the lengths of the NPC hand splines so that there is less clipping with the bar.

Made a silly mistake last night when moving where we set the in game bool in NPCs in the main game managers as also moved where we set game active in the game managers so moved setting game active back to where it should be so now all the games can be properly started again.