Changed how doing the held accessories for animations so easier to understand and use in future. Changed where we update in game for finishing the games to be the same as where we set the game not active anymore which is after a 4 second timer. Clearing held and placed accessories when updating in game.
New array of static mesh components in NPC.
This is filled on begin play with 0 index always being accessory 1 and 1 index always being accessory 2.
Then changed all associated anim notifies to use this new system making use of a new interface function in NPC interface for getting this array.
This just makes it way easier to identify which accessory is which and I guess optimises it a little as well since not doing component tag checks and selecting instead just using the 0 and 1 indexes of this array.

In all game managers where we originally call update in game function in the NPC opponents for finishing the game we now call it when we reset in the game active bool for the game manager which is in a separate function called with a 4 second timer, this means that the animation they were going to do originally if it was a character specific action only resumes when they are returned to their neutral position.

Added functionality for the update in game event in the NPCs where they now clear the held accessories and placed accessories as well.