Drink order interface event and changing drink mesh in drink making animation for the drink order index. Jukebox fix for starting song title and current song index. New anim notify for ending character specific actions. Placed and held accessories not affected by colourful lighting like the cards. Placing accessory using trace for placing directly onto surface. Changed pivots of all cocktails and shakers.
New interface event for getting the drink order index from the given NPC.
Using this in the equip accessory notify for changing the mesh of the glass that is being filled / served.

Jukebox fixed for setting the current song index properly on start and also setting the track title properly since using random song starting.

New anim notify which disables the character specific action bool in the NPC. Think might have made a new interface event specifically to do this cant really remember tbh.

Placed and held accessories now only affected by lighting channel 1. Had to set this explicitly in the dequip notify when placing the accessory.

Changed a lot of the sockets on skeleton for the drink making animation.
Changed the pivots on all cocktails and the shaker.
New functionality in the dequip notify using line trace to determine where to place the accessory on the nearest surface below it so had to change the pivots so that they are not phasing in the surface.