Simplified hovering and unhovering for rules.
Now using get current animation time to determine when to start animation from for hovering and unhovering in the rules widget.
New button in rules widget which is invisible just using it for the hovering and unhovering events.
Its kinda fixed most of the visual issues you can still glitch it out a bit but think its less so now.