Biel Animations. Added functionality in start game indicator for NPCs specifically for biel. Moved Biel positions for rat screw and neutral and moved card piles for rat screw. Rat screw dialogue activation when slapping. NPC audio component volume multiplier set based on NPC name. When updating in game also stopping character specific action. Advanced calculate next int anim notify.
Biel animations added and added to animation blueprint.
New entries in character specific action enum for these new animations since Tino doesn't want to do animations the same way everyone else should be doing them.

Moved the positions for neutral and rat screw opponent position.
Changed player position for rat screw.

Rat screw more do dialogue calls for the slapping both player and opponent.

Calculate next state int anim notify updated to use array of states so can do multiple with one notify.
Use this for in-game animations that are shared between this and idle animations.

Dante actually uses his idle animations now properly in game as well.

Altered the timings for the start game fade in and fade out as well as actually activating the game so that it is longer so that there is more time for the game invitation dialogue.