Linking dialogue almost all done just waiting on the rest of biel dialogue animations.
Link anim notifies in the linking between zilia and dante. However need some more animations for Biel to do the rest of the linking dialogue.
Now need to do functionality for activating these linking dialogues since they are kinda just random and mostly when the NPCs are not in a game.