Schrodingers cat automatically moving to surface using root location so always on something and not in the air. Dante update lip sync animations imported.
New functionality for moving the cat where we set the skeletal mesh directly onto the surface below it precisely using the root bone of the skeletal mesh which means that even if we set the target point to be in the air then the mesh will be directly on the surface below it without any clipping.

Added the updated dante lip sync animations thankfully didn't have to do all the anim notifies again so didn't crashout yay.