Fixed some errors in the main level. Changed player walk speed to be 350 from 600.
Had to delete old Biel animation that was still in Ryn's version of the game but not the main. The animation is still being used just in the NPC animation blueprint instead of in the world.
Also had to delete a Group actor data thing from the main level but don't think this should change anything other than get rid of the error.
Had to set two bottles in the main corridor to not simulate physics since their collisions are not setup for it and they were producing warnings also probably shouldn't be simulating physics on small objects since they can cause unexpected outcomes with interacting with player movement and mesh.

Lowered the player walk speed to 350 from 600 so we are not as fast at movement since we were moving around the environment a bit too fast.