Cat no longer makes noise or looks at player when they are sleeping. Updating character to look at and scan cards when in main games depending on who's turn it is and drawing. More biel dialogue animations added. Conversation manager change. Look at and scan card activations in main games. Zilia new animations.
New checks for making noise and looking at player when the current cat stationary animation is the sleeping ones.

Added more of biel's dialogue animations to the animation blueprint and setup their notifies.

Changed the conversation manager to not now reset the speaker's dialogue incase it is the game invitation one which should override. Also now if dialogue is active when it tries to start a conversation then it will activate the check loop again after 2 seconds.

In main games changing and turn manager updating whether NPCs will scan their cards or look at another character depending if it is their turn. In auction they will look at a random drawing character when the round is a draw.

Zilia idle and leg cross animations added and added to animation blueprint.