Moved darts playing and non-playing positions. Moved chair near darts playing position. No more Tuna eye tracking so no more bugging with the eye's being weird whilst doing noise animation.
Moved the darts playing and non-playing positions a little in the main level.
Moved the stool that was next to the playing position a bit more underneath the table.

Didn't remove the eye look at functionality for Tuna but have not got it being used in the animation blueprint since this was bugging when Tuna would blend for the make noise animation so no more bugging but also no more eye tracking.