Animation conversation fixing. Cat fixing. Conversation manager fixes. Cat custom rotation depending on stationary animation when moving.
Some fixes for animation conversations with anim notifies. Need to properly go through each one and check them all but going to do that after Tino gives me the final lip sync animations.

Conversation manager was calling function with timer that just wasn't the right name so now thats fixed.

Cat skeletal mesh local rotation set custom to the stationary animation it is going to loop through so that they are more facing the direction of the target point when moving.