Moved poker chips on bar to more empty space rather than play area. Leading cat update. Updated song titles.
Leading cat more functionality for moving the skeletal mesh of the cat to be directly where it should be on the ground as well as its rotation using the normal of the line trace.
Changed the length of time for the leading cat to be longer so it moves a little slower since player moves slower now.
Just need the cat moving animation now for this to be finished.
Changed starting position of the leading cat.
Audio component for leading cat is now attached to the cat skeletal mesh.

Updated song titles in files and enum for track titles:
3 - When Is It Over
4 - Nightfall's Radiance