More culling. Sound cue creation. Adding sounds to actual functionality. Changed when playing some sounds like not playing on unhovering button. Pause menu animations 60fps instead of 20 might be better or might be placebo tbh.
Created sound cues for various things.
Added a lot of the audio in the game with the sound cues.
Unhovering for buttons dont play sounds now.
Pause menu animations now at 60fps since they looked really low frame rate but dont know if this changed anything or if it was placebo.