Tuna stretch animation added and functionality. Rules hovering and unhovering sounds.
New state machine in the Tuna animation blueprint for doing the stretch animation and then doing the moving animation.
New functionality for making Tuna stretch the first time using a new variable and a do once.
Now when hovering and unhovering the rules texts we play the card move sound 2d.