Reset game button. More dialogue linking. Play animation sound notify was reset with no sound cues so had to update them all. Dialogue update and finialisation for game invitation and dialogue linking.
New debug button which is an actual input action so can be used in shipped version of the game that just resets the game the same way the game completed reset button does.

Play animation sound notify instances were reset resulting in no sound from them so changed these back. Will need to check these again when going to build to make sure they are right for the built version.

Dialogue updates and changes and everything should be working as intended now according to the script.
You can still interrupt the dialogue chains with starting a game but thats staying in since the player is the centre of the game and otherwise we're going to have more bugs anyway so yea.