Filling save data structs. NPC game info functionality setup in NPC.
Populated the save data structs with what I think is necessary info for the save games.
New interface for the game info widget displayed next to NPCs for setting the NPC ref, updating the info and playing the pop-up animation forward and reverse.
These are called in the NPC player interaction graph.