More save data functionality.
More changes to the structs for saving.
Load and save functions in each game manager.
Almost all of those are setup but not tested.
Will continue working on it so can quit a game and then save the data and then load that data then back in when starting the game again.