More save system loading and saving functionality in game managers. End game functions for each game manager which are called when not finishing a game but exiting a game. New functionality in game mode for game complete but not game over.
Needs more testing but think it is working now so should be able to leave the games during their run and then start them again from the same point.
Need to also test the saving of the data to text file and reading it from text file but will do that later.