IDI locking camera and focusing camera on character when new player turn.
New event in the IDI game manager.
Sets ignore look input.
Then finds the character who's turn it is now and if they are not the player then we focus the camera on them using a timeline to lerp the rotator to the find look at rotation.

For loading functions in the game managers we now check if we have loaded and if not then we set all the variables to their default.

For now set the start game timers for the NPCs to be like 3 seconds since the dialogue is really annoying me.