Removed all casting for the game mode in place of using an interface now. IDI focus camera update for doubting and when finishing doubting showcase.
No more hard references I think to the game mode now from other blueprints just using a new interface for it.

IDI focus camera now has a actor input to specify from outside calls what actor to focus and so we can then use this to focus the main pile when it is the player's turn and also focus on the character calling doubt if it is not the player.