Difficulties overhaul. More saving fixes. More game mode interface fixes.
I doubt it and auction difficulties overhaul for what cards are given to players and are used in the game so that they more scale with complexity as well as actual difficulty now.
IDI easy: 3 players - no unique cards - 30 cards

IDI medium: 4 players - 4 unique cards - 40 cards

IDI hard: 4 players - 8 unique cards - 60 cards

Auction Easy: 4 cards auctioning - 3 players - no unique cards

Auction Medium: 6 cards auctioning - 4 players - 4 unique cards

Auction Hard: 8 cards auctioning - 4 players - 8 unique cards

Changes to the game mode, game managers for auction and IDI and changes to the card dealer as well as changes to the initiate game function in the NPC.
Saving more info now for the NPC difficulties and names so that they will save after the player completes a game and the difficulties increase.
Changes to the scalar difficulty function in the game mode to always increase difficulty if the NPC is the owner of the game (if they have the corresponding game name).

Fixed some mistakes with some of the get functions in the game mode which I didn't fully finish when doing the interface stuff yesterday.