Fixed rules spamming glitchyness.
New bool in rules manager for can show rules.
Set to false when hiding rules.
Set as true for default.
New function in rules manager for setting this bool back to true after a short delay using a set timer by function name.
This gets rid of the glitchyness when spamming toggle rules but means that there is a very small delay between being able to show and hide the rules but better than the visual glitchyness.