Rules setup based on difficulty. Specific unique cards for medium difficulties in auction and IDI and rat screw. Rules changing unique card display based on difficulty. NPC info display visual update and auto updating info when visible.
New function in the rules widget interface which passes the difficulty and then changes the displayed rules based on this.
This function is called in the show rules function in the rules manager.

New functionality in card dealer for specific unique cards being added for the medium difficulties of all games in the card dealer in the setup cards and deal cards functions.

The rules unique card display now changes based on the difficulty since there are specific unique cards now that are used in the medium difficulty of each game.
This also updates the positioning for the easy difficulty rules since there are only 2 rules cards on screen.

The NPC info display visual update with the scroll thingy on the text box. Also now when it is visible we are auto updating the info every 0.25 seconds so that if there happens to be changes to the NPC's difficulty it is represented here as well. The timer is cleared when the info is hidden again.